fixup PR #43

+++ src/SoundExporter.h
@@ -20,7 +20,7 @@ public:
 
     void Export(const std::vector<SongEntry>& entries);
 private:
-    static void writeSilence(SNDFILE *ofile, float seconds);
+    static void writeSilence(SNDFILE *ofile, double seconds);
     size_t exportSong(const std::filesystem::path& fileName, uint16_t uid);
 
     SongTable& songTable;
